API version 2.0

+++ src/TelegramBot/types/location.cr
@@ -3,8 +3,8 @@ require "json"
 module TelegramBot
   class Location
     JSON.mapping({
-      longitude: Int32,
-      latitude:  Int32,
+      longitude: Float64,
+      latitude:  Float64,
     })
   end
 end
